base: avoid implicit conversions
This patch is a prerequisite for compiling the code with
the warnings -Wconversion enabled.

Issue #23

+++ repos/base-foc/src/core/signal_source_component.cc
@@ -41,7 +41,7 @@ void Signal_source_component::submit(Signal_context_component &context,
                                      unsigned long             cnt)
 {
 	/* enqueue signal to context */
-	context.increment_signal_cnt(cnt);
+	context.increment_signal_cnt((int)cnt);
 
 	if (!context.enqueued()) {
 		_signal_queue.enqueue(context);
