+++ repos/base-nova/src/lib/base/ipc.cc
@@ -42,7 +42,7 @@ Rpc_exception_code Genode::ipc_call(Native_capability dst,
 		/* calculate max order of caps to be received during reply */
 		unsigned short log2_max = 0;
 		if (rcv_caps) {
-			log2_max = log2(rcv_caps);
+			log2_max = (uint16_t)log2(rcv_caps);
 
 			/* if this happens, the call is bogus and invalid */
 			if ((log2_max >= sizeof(rcv_caps) * 8))
@@ -93,7 +93,7 @@ Rpc_exception_code Genode::ipc_call(Native_capability dst,
 	if (utcb.msg_words() < 1)
 		return Rpc_exception_code(Rpc_exception_code::INVALID_OBJECT);
 
-	return Rpc_exception_code(copy_utcb_to_msgbuf(utcb, rcv_window, rcv_msg));
+	return Rpc_exception_code((int)copy_utcb_to_msgbuf(utcb, rcv_window, rcv_msg));
 }
