+++ repos/base/include/base/allocator_avl.h
@@ -279,7 +279,7 @@ class Genode::Allocator_avl_base : public Range_allocator
 
 		Alloc_result try_alloc(size_t size) override
 		{
-			return Allocator_avl_base::alloc_aligned(size, log2(sizeof(addr_t)));
+			return Allocator_avl_base::alloc_aligned(size, (unsigned)log2(sizeof(addr_t)));
 		}
 
 		void free(void *addr, size_t) override { free(addr); }
