+++ repos/base/src/core/include/trace/root.h
@@ -38,7 +38,7 @@ class Genode::Trace::Root : public Genode::Root_component<Session_component>
 		{
 			size_t ram_quota       = Arg_string::find_arg(args, "ram_quota").ulong_value(0);
 			size_t arg_buffer_size = Arg_string::find_arg(args, "arg_buffer_size").ulong_value(0);
-			unsigned parent_levels = Arg_string::find_arg(args, "parent_levels").ulong_value(0);
+			unsigned parent_levels = (unsigned)Arg_string::find_arg(args, "parent_levels").ulong_value(0);
 
 			if (arg_buffer_size > ram_quota)
 				throw Service_denied();
